+++ src/main/java/com/projetosjavaUFBA/javaProject/services/TeacherService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.projetosjavaUFBA.javaProject.entities.Teacher;
 import com.projetosjavaUFBA.javaProject.repositories.TeacherRepository;
+import com.projetosjavaUFBA.javaProject.services.exceptions.DatabaseException;
 import com.projetosjavaUFBA.javaProject.services.exceptions.ResourceNotFoundException;
 @Service
 public class TeacherService {
@@ -28,8 +30,16 @@ public class TeacherService {
 		return teacherRepository.save(obj);
 	}
 	public void delete(Long id) {
-		teacherRepository.deleteById(id);
-	}
+	    try {
+	        if (teacherRepository.existsById(id)) {
+	        	teacherRepository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    } catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseException(e.getMessage());		
+	    }	
+	} 
 	public Teacher update(Long id, Teacher obj) {
 		Teacher entity = teacherRepository.getReferenceById(id);
 		updateData(entity,obj);
